+++ r-kvstore-20150101/alibabacloud_r_kvstore20150101/models/_switch_instance_to_target_zone_response_body.py
@@ -8,8 +8,10 @@ class SwitchInstanceToTargetZoneResponseBody(DaraModel):
     def __init__(
         self,
         request_id: str = None,
+        task_id: str = None,
     ):
         self.request_id = request_id
+        self.task_id = task_id
 
     def validate(self):
         pass
@@ -22,12 +24,18 @@ class SwitchInstanceToTargetZoneResponseBody(DaraModel):
         if self.request_id is not None:
             result['RequestId'] = self.request_id
 
+        if self.task_id is not None:
+            result['TaskId'] = self.task_id
+
         return result
 
     def from_map(self, m: dict = None):
         m = m or dict()
         if m.get('RequestId') is not None:
             self.request_id = m.get('RequestId')
+
+        if m.get('TaskId') is not None:
+            self.task_id = m.get('TaskId')
 
         return self